Added a new class file for the 'Job' class.
