+++ src/Main.java
@@ -3,82 +3,6 @@ import java.util.List;
 import java.util.Random;
 
 /**
- * Job is a set of operations to be taken in particular order.
- * 
- * @author Todor Balabanov
- */
-class Job {
-	/**
-	 * Job title.
-	 */
-	private String name = "";
-
-	// TODO May be LinkedList is better choice for this member field.
-	/**
-	 * List of operations taken for this job.
-	 */
-	private List<Operation> operations = new ArrayList<Operation>();
-
-	/**
-	 * Constructor with all parameters.
-	 * 
-	 * @param name
-	 *            Job name.
-	 */
-	public Job(String name) {
-		this.name = name;
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#toString()
-	 */
-	@Override
-	public String toString() {
-		return "Job [name=" + name + ", operations=" + operations + "]";
-	}
-
-	/**
-	 * Name field getter.
-	 * 
-	 * @return Job name.
-	 */
-	public String getName() {
-		return name;
-	}
-
-	/**
-	 * Name field setter.
-	 * 
-	 * @param name
-	 *            Job name.
-	 */
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	/**
-	 * Operation list setter.
-	 * 
-	 * @param operations
-	 *            The list of operations.
-	 */
-	public void setOperations(List<Operation> operations) {
-		this.operations = operations;
-	}
-
-	/**
-	 * Operation list getter.
-	 * 
-	 * @return The list of operations.
-	 */
-	public List<Operation> getOperations() {
-		return operations;
-	}
-}
-
-/**
  * Application single entry point class.
  * 
  * @author Todor Balabanov
